Eliminated prepared statements (by default).  Requires rhubarb 0.2.5.
wallaby-agent will not use prepared statements unless the env contains
WALLABY_USE_PREPARED_STATEMENTS=1 in the environment.  We've seen some
unusual behavior with prepared statements in sqlite 3.3.6-5 and
ruby-sqlite3 1.2.4-1.  In environments with more recent sqlite
libraries (e.g. on F12), for example, it will be faster to run with
prepared statements enabled.