Support for versioned configurations.
This is distinct from snapshot support and enables wallaby clients to
request previous versions of a node's configuration.  This supports
concurrent modification of the store by multiple clients without
waiting for configurations to be pushed out; it is also necessary in
order to support "implicit subsystems."  Configuration versions are
stored in the snapshot database.

There are two implementations of versioned configurations: a
"normalized" implementation and a "serialized" implementation.  The
serialized implementation, which simply stores gzipped serialized
dictionaries to the database, is enabled by default.  The "normalized"
implementation is able to exploit redundancy between the configs of
multiple nodes; it is likely to be slower but more space-efficient.
(The normalized implementation is currently untested.)

User-visible changes with this commit:

*  Node#getConfig now takes a hash of options.  The only supported
   option for now is "version", which takes a version as an int and
   will return the latest configuration for that node that is less
   than or equal to that version.  (Versions are intended to be opaque
   but currently correspond to UTC timestamps from the store.)

*  This functionality requires a db migration to version 5.