Wallaby dump files no longer contain (empty) per-feature subsystem lists.
We dropped support for explicit subsystems recently but had retained
it in the dump format for backwards compatibility with dumps from
earlier in the development cycle.  However, the Ruby YAML library will
ignore extraneous fields specified in a serialized object, so there is
no harm in removing the explicit subsystem lists entirely.