Proactive feature inconsistency avoidance (related to BZ 614442).
Updates to feature inclusions, dependencies, and conflicts are checked
to ensure that (1) the inclusion/dependency graph is acyclic (i.e. so
that F cannot transitively include or depend upon itself), and (2) no
feature that transitively includes or depends on some feature F can
also have a (potentially transitively-carried) conflict with F.  There
are minor performance implications, since we solve for the transitive
closure of the inclusion/dependency graph after each feature
operation; this is cubic in the number of features that have either
incoming or outgoing inclusion or dependency edges.

This is a squash merge of the following commits:

commit af4da2f4d045ee803a8fa0d4b72a056ac0ab9488
Author: Will Benton <willb@redhat.com>
Date:   Sun Aug 1 14:48:02 2010 -0500

commit 3021e9d4a03d2b3b865bea08c0966f10c2df130a
Author: Will Benton <willb@redhat.com>
Date:   Sat Jul 31 10:50:02 2010 -0500

commit 045516dce4d2ad8125047a919140011f53dd1099
Author: Will Benton <willb@redhat.com>
Date:   Sat Jul 31 01:44:37 2010 -0500

commit 5f77a4aae4f0a34994d3f44ee2b19c24d568a915
Author: Will Benton <willb@redhat.com>
Date:   Sat Jul 31 00:24:00 2010 -0500

commit f571c0c4ede578bcf4f364308ffdf22ffa06f6a8
Author: Will Benton <willb@redhat.com>
Date:   Fri Jul 30 17:33:14 2010 -0500

commit e9e1c5787be49c9d1ba9b2f110570c68601d691f
Author: Will Benton <willb@redhat.com>
Date:   Thu Jul 29 17:13:21 2010 -0500