Removed QMF console from wallaby-agent.
w-a had a console to ensure that two w-a processes could not run
against the same broker.  However, because the act of checking for
another w-a and starting up a w-a is not atomic, this introduced
a rare but unavoidable race.  Since all wallaby agents will have
the same agent label, at most one can exist on any broker.  Because
the added console did not introduce any safety (due to the race
and to the fact that the broker will prevent multiple wallaby agents
from coexisting, atomically) but did introduce performance problems
(due to the console having to handle management events), it was
a good candidate for removal.

I am reasonably confident that pmackinn's original observation of
two w-as on the same broker was due to an older version of
wallaby-agent (with a different agent label) running at the same time
as an agent with the current label.