Added several command-line options to HTTP server.
The HTTP server now runs as a daemon by default, and drops
privileges if it is run as root.  Use the -f or --foreground
option to run in the foreground, and the --run-as USER option
to specify a user to run as.

If running in the foreground, you can specify -q or --quiet
to disable logging requests to stderr.