cmd_shebang support
cmd_shebang is more suitable for general-purpose scripts than
cmd_console; it allows users to pass arguments to their scripts.

This commit also includes a new Wallaby shell command callback:
:preprocess_options.  A :preprocess_options callback must take a
list of command-line arguments and return a list of command-line
arguments (after some transformation).  These will be called
in the order they are declared (for now; do not depend on this)
and the result of applying all of them to the argument list will
be the command-line parameters passed to the :before_option_parsing
callback and the option parser.