API and tool improvements; API version 20101031.5
Annotations are now at least partially supported in the command line tools.

Adding an annotation to an entity now dirties that entity.

The list of entities whose configuration or metadata have changed since the
last configuration activation is available via the Store#affectedEntities
method.

The list of nodes that will be affected by the next configuration activation is
available via the Store#affectedNodes API method.