More resilient activation on F17.
This is both a workaround for a problem with QMF on F17 (viz.,
we can't send events) and a general improvement to configuration
activation.  Specifically, configuration events are a performance
optimization but are not required for correctness.  Before this
commit, failure to send a configuration event would result in failure
of the Store#activateConfiguration API method.  (It would also leave
node metadata in an inconsistent state.)  Now, failure to send a
configuration event merely results in a warning.