Spurious configuration removal off by default.
The Wallaby agent no longer removes possibly-spurious versioned
configurations on startup.  (This is a cleanup measure from a bug
in a very old version of Wallaby; it is unlikely necessary any more.)
To enable this behavior, set WALLABY_REMOVE_SPURIOUS_CONFIGS to yes
or true in the environment in which wallaby-agent runs.